+++ server/controllers/index.js
@@ -1,0 +1,24 @@
+const { showProjects, showProject, searchProjects, searchCategory } = require('./projects');
+const { showAboutInformation } = require('./about');
+const { showProfile } = require('./profile');
+const { showCommentaryForm, createCommentary } = require('./commentary');
+const { showAuthForm, register, authorization, logout } = require('./auth');
+
+module.exports = {
+  setControllerOperation(param) {
+    return {
+      showProjects,
+      showProject,
+      searchCategory,
+      searchProjects,
+      showAboutInformation,
+      showProfile,
+      showAuthForm: showAuthForm(param),
+      register,
+      authorization,
+      logout,
+      showCommentaryForm,
+      createCommentary
+    };
+  }
+};
