--- server/middlewares/auth.js
@@ -1,30 +1,0 @@
-const modelUser = require('../models/users');
-
-module.exports = {
-    findUser(req, res, next) {
-        if (req.session) {
-            modelUser.findById(req.session.userId)
-                .then(user => {
-                    req.user = user;
-                    res.locals.user = user;
-
-                    next();
-                })
-                .catch();
-        } else {
-            next();
-        }
-    },
-
-    authenticated(req, res, next) {
-        if (req.user) return next();
-
-        res.status(403).redirect('/auth/login');
-    },
-
-    unauthenticated(req, res, next) {
-        if (!req.user) return next();
-
-        res.redirect('/books');
-    }
-};
