+++ server/models/about.js
@@ -5,8 +5,8 @@ const Schema = mongoose.Schema;
 
 const aboutCollection = new Schema({
   title: String,
-  description: String,
-},{collection:'about'});
+  description: String
+}, { collection: 'about' });
 
 const modelAbout = mongoose.model('about', aboutCollection);
